fix(metrics): Redis::pconnect() 포트에 문자열 넘기던 것 캐스팅
REDIS_HOST를 이제 environment.ini에서 실제로 읽어오게 되면서(직전 커밋)
드러난 문제: REDIS_PORT도 ini값이라 항상 문자열인데, 이 Redis extension
버전은 pconnect() 두번째 인자로 엄격하게 int를 요구해서 "expects parameter 2
to be integer, string given"으로 매번 실패하고 있었음. (int) 캐스팅 추가.

+++ app/Lib/Metrics.php
@@ -92,7 +92,7 @@ class Metrics {
                 try {
                     if($redis->pconnect(
                         Config::getEnvironmentData('REDIS_HOST'),
-                        Config::getEnvironmentData('REDIS_PORT') ? : 6379,
+                        (int)(Config::getEnvironmentData('REDIS_PORT') ? : 6379),
                         Config::REDIS_OPT_TIMEOUT
                     )){
                         if($auth = Config::getEnvironmentData('REDIS_AUTH')){
